Modify update methods for User and ConfirmationToken
