+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/config/JwtService.java
@@ -7,6 +7,7 @@ import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 
 import java.security.Key;
 import java.util.Date;
