+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/config/SecurityConfiguration.java
@@ -10,26 +10,37 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 @RequiredArgsConstructor
 @EnableWebSecurity
-public class SecurityConfiguration {
+public class SecurityConfiguration implements WebMvcConfigurer {
 
     private final JwtAuthenticationFilter jwtAuthFilter;
     private final AuthenticationProvider authenticationProvider;
 
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:8080", "http://localhost:4200")
+                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
+                .allowedHeaders("*")
+                .allowCredentials(true)
+                .maxAge(3600);
+    }
+
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+
         http
                 .csrf()
                 .disable()
                 .authorizeHttpRequests()
-                .requestMatchers("**")
-                //.requestMatchers("/api/v1/auth/**")
-//                .permitAll()
-//                .requestMatchers("/api/v1/auth/authenticate")
-                .permitAll()
+                .requestMatchers("**").permitAll()
+                .requestMatchers("/api/v1/auth/**").permitAll()
+                .requestMatchers("/api/v1/pwdres/**").permitAll()
                 .anyRequest()
                 .authenticated()
                 .and()
@@ -38,7 +49,11 @@ public class SecurityConfiguration {
                 .and()
                 .authenticationProvider(authenticationProvider)
                 .addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class)
-                .headers().frameOptions().disable();
+                .formLogin();
+//                .headers().frameOptions().disable();
+
+        http.headers().frameOptions().disable();
+        http.cors().and().csrf().disable();
 
         return http.build();
     }
