+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/controllers/RegistrationController.java
@@ -1,19 +1,21 @@
 package com.nesslabs.nesslabspring.controllers;
 
 import com.nesslabs.nesslabspring.dto.RegistrationRequest;
-import com.nesslabs.nesslabspring.repositories.ConfirmationTokenRepository;
 import com.nesslabs.nesslabspring.services.RegistrationService;
-import com.nesslabs.nesslabspring.services.UserService;
 import com.nesslabs.nesslabspring.exception.InvalidConfirmationEmail;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping(path = "api/v1")
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class RegistrationController {
+
+    @Autowired
     private final RegistrationService registrationService;
 
     @PostMapping("/auth/registration")
@@ -21,12 +23,10 @@ public class RegistrationController {
         try {
             registrationService.register(request);
         }catch (InvalidConfirmationEmail e){
-            return ResponseEntity.badRequest().body(e.getMessage());
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         }
         return ResponseEntity.status(HttpStatus.CREATED). body("One more step.Please verify your account");
-
     }
-
 
     @GetMapping("/auth/registration/confirm")
     ResponseEntity<String> confirm(@RequestParam("token") String token) throws InvalidConfirmationEmail {
