+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/repositories/ConfirmationTokenRepository.java
@@ -2,33 +2,16 @@ package com.nesslabs.nesslabspring.repositories;
 
 import com.nesslabs.nesslabspring.entity.ConfirmationToken;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Repository
 public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
 
 
-    @Query("SELECT c from ConfirmationToken c where c.token=:token ")
     Optional<ConfirmationToken> findByToken(@Param("token") String token);
 
-    @Transactional
-    @Modifying
-    @Query("UPDATE ConfirmationToken c " +
-            "SET c.confirmedAt = ?2 " +
-            "WHERE c.token = ?1")
-    int updateConfirmedAt(String token,
-                          LocalDateTime confirmedAt);
-
-
-    @Transactional
-    void deleteByuser_Id(int table1Id);
-
+    ConfirmationToken findByuser_Id(int table1Id);
 }
