+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/services/ConfirmationTokenService.java
@@ -1,38 +1,12 @@
 package com.nesslabs.nesslabspring.services;
 
 import com.nesslabs.nesslabspring.entity.ConfirmationToken;
-import com.nesslabs.nesslabspring.repositories.ConfirmationTokenRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
+public interface ConfirmationTokenService {
 
+    void saveConfirmationToken(ConfirmationToken token);
 
-@Service
-public class ConfirmationTokenService {
+    ConfirmationToken getToken(String token);
 
-    private final ConfirmationTokenRepository confirmationTokenRepository;
-
-    @Autowired
-    public ConfirmationTokenService(ConfirmationTokenRepository confirmationTokenRepository) {
-        this.confirmationTokenRepository = confirmationTokenRepository;
-    }
-
-    public void saveConfirmationToken(ConfirmationToken token){
-        confirmationTokenRepository.save(token);
-    }
-
-    public ConfirmationToken getToken(String token) {
-        return confirmationTokenRepository.findByToken(token).orElse(null);
-    }
-
-    public int setConfirmedAt(String token) {
-        return confirmationTokenRepository.updateConfirmedAt(
-                token, LocalDateTime.now());
-    }
-
-    public void deleteToken(int id){
-        confirmationTokenRepository.deleteByuser_Id(id);
-    }
-
+    ConfirmationToken findTokenByUserId(int id);
 }
