+++ ness-lab-spring/src/main/java/com/nesslabs/nesslabspring/services/ConfirmationTokenServiceImpl.java
@@ -1,0 +1,31 @@
+package com.nesslabs.nesslabspring.services;
+
+import com.nesslabs.nesslabspring.entity.ConfirmationToken;
+import com.nesslabs.nesslabspring.repositories.ConfirmationTokenRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class ConfirmationTokenServiceImpl implements ConfirmationTokenService{
+
+    private final ConfirmationTokenRepository confirmationTokenRepository;
+
+    @Autowired
+    public ConfirmationTokenServiceImpl(ConfirmationTokenRepository confirmationTokenRepository) {
+        this.confirmationTokenRepository = confirmationTokenRepository;
+    }
+
+    public void saveConfirmationToken(ConfirmationToken token){
+        confirmationTokenRepository.save(token);
+    }
+
+    public ConfirmationToken getToken(String token) {
+        return confirmationTokenRepository.findByToken(token).orElse(null);
+    }
+
+    public ConfirmationToken findTokenByUserId(int id){
+        return confirmationTokenRepository.findByuser_Id(id);
+    }
+
+}
